Organize CCR adapters and abstract daily attendance as adapter
