+++ app/services/ccr/daily_attendance_adapter.rb
@@ -1,0 +1,48 @@
+# CCR requires the total number of daily attendances.
+# This can be derived from CCCD data as follows:
+# The number of attendances claimed is equal to
+# the total of quantities claimed for the 3 DAF/H/J
+# basic fees + 2 (included in the basic fee). If there
+# are none claimed then the attendance can be said to be
+# the actual trial length
+module CCR
+  class DailyAttendanceAdapter
+    attr_reader :claim
+
+    def initialize claim
+      @claim = claim
+    end
+
+    class << self
+      def attendances_for claim
+        adapter = new(claim)
+        adapter.attendances
+      end
+    end
+
+    def attendances
+      if daily_attendance_uplifts?
+        daily_attendance_uplifts + DAILY_ATTENDANCES_IN_BASIC
+      else
+        [claim.actual_trial_length, DAILY_ATTENDANCES_IN_BASIC].compact.min
+      end
+    end
+
+    private
+
+    # The first 2 daily attendances are included in the Basic Fee (BAF)
+    DAILY_ATTENDANCES_IN_BASIC = 2.freeze
+
+    def daily_attendance_fee_types
+      ::Fee::BasicFeeType.where(unique_code: %w[BADAF BADAH BADAJ])
+    end
+
+    def daily_attendance_uplifts
+      @attendance_uplifts ||= claim.fees.where(fee_type_id: daily_attendance_fee_types).sum(:quantity).to_i
+    end
+
+    def daily_attendance_uplifts?
+      daily_attendance_uplifts > 0
+    end
+  end
+end
